+++ client/src/components/modals/index.ts
@@ -1,0 +1,2 @@
+export { ModalAdmin } from './modal-admin';
+export { ModalContent } from './modal-content';
