+++ client/src/components/modals/modal-content/styles.ts
